Utility file to Utility Class [PT2]
This is a pretty big looking change set, but it’s really not.

Essentially I moved all of the Utility file into a new namespace of Contentstack\Support and defined all the methods within the Utility class as static. This allows these methods to be autoloader and keeps the functionality similar to the current.