Fix 2 issues with language method
The language method sets a default of an empty string. This should be ok based on the docs - however the helper method checks for empty strings and throws exceptions. Because of this we should set the default to an actual value.

Based on the comments the default should be ‘en-us’ so I’m using that here as well.